requirements.txt
aiogram==3.4.1
python-dotenv
requests
pydantic==2.5.3